switch case, moving state classes, const expr
